python/multi_agentic_claude/requirements.txt
# semantic-kernel==1.3.0
# openai>=1.0.0
# httpx>=0.27.0
# pydantic>=2.0.0
# python-dotenv>=1.0.0
# python-dateutil>=2.9.0

# # Testing
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=5.0.0
# pytest-mock>=3.14.0
# respx>=0.21.0


# Core — use latest SK which is fully pydantic v2 compatible
semantic-kernel>=1.39.0

# Pin pydantic to a version known-good with SK 1.39.x
pydantic>=2.10.4,<3.0.0
pydantic-settings>=2.4.0

# Other deps
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
respx>=0.21.0